module_2(fix): refactor imports
