+++ src/components/home-page-blocks/news/buttons-block/buttons-block.module.scss
@@ -1,4 +1,4 @@
-@use '@styles/colors';
+@use "@styles/colors";
 
 .buttons__wrapper {
   display: flex;
