+++ src/components/home-page-blocks/subscription/subscription-form/index.ts
@@ -1,0 +1,1 @@
+export * from "./subscription-form";
