+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import { App } from "./components/app/App.tsx";
+import { App } from "@components/app/App.tsx";
 import "./main.css";
 
 createRoot(document.getElementById("root")!).render(
