--- src/pages/main-layout-page/index.ts
@@ -1,1 +1,0 @@
-export * from "./main-layout-page";
