+++ src/styles/_mixins.scss
@@ -1,4 +1,4 @@
-@use "./media";
+@use "@styles/media";
 
 @mixin mainLayout {
   width: 90%;
