+++ src/ui/icons/arrow-right/arrow-right.tsx
@@ -1,4 +1,4 @@
-import { IconsWrapper } from "../icons-wrapper/icons-wrapper";
+import { IconsWrapper } from "../icons-wrapper";
 
 export const ArrowRight = () => {
   return (
