+++ src/ui/icons/index.ts
@@ -1,3 +1,3 @@
 export { ArrowLeft } from "./arrow-left/arrow-left";
 export { ArrowRight } from "./arrow-right/arrow-right";
-export { IconsWrapper } from "./icons-wrapper/icons-wrapper";
+export { IconsWrapper } from "./icons-wrapper";
